+++ src/tokens/user_token.rs
@@ -18,11 +18,12 @@ pub struct UserToken {
     pub access_token: AccessToken,
     client_id: ClientId,
     client_secret: Option<ClientSecret>,
-    login: Option<String>,
+    /// Username of user associated with this token
+    pub login: String,
     /// The refresh token used to extend the life of this user token
     pub refresh_token: Option<RefreshToken>,
     /// Expiration from when the response was generated.
-    expires_in: Option<std::time::Duration>,
+    expires_in: std::time::Duration,
     /// When this struct was created, not when token was created.
     struct_created: std::time::Instant,
     scopes: Vec<Scope>,
@@ -30,12 +31,14 @@ pub struct UserToken {
 
 impl UserToken {
     /// Assemble token without checks.
+    ///
+    /// If `expires_in` is `None`, we'll assume `token.is_elapsed() == true`
     pub fn from_existing_unchecked(
         access_token: impl Into<AccessToken>,
         refresh_token: impl Into<Option<RefreshToken>>,
         client_id: impl Into<ClientId>,
         client_secret: impl Into<Option<ClientSecret>>,
-        login: Option<String>,
+        login: String,
         scopes: Option<Vec<Scope>>,
         expires_in: Option<std::time::Duration>,
     ) -> UserToken {
@@ -45,9 +48,9 @@ impl UserToken {
             client_secret: client_secret.into(),
             login,
             refresh_token: refresh_token.into(),
-            expires_in,
+            expires_in: expires_in.unwrap_or_default(),
             struct_created: std::time::Instant::now(),
-            scopes: scopes.unwrap_or_else(Vec::new),
+            scopes: scopes.unwrap_or_default(),
         }
     }
 
@@ -69,9 +72,9 @@ impl UserToken {
             refresh_token.into(),
             validated.client_id,
             client_secret,
-            validated.login,
+            validated.login.ok_or(ValidationError::NoLogin)?,
             validated.scopes,
-            validated.expires_in,
+            Some(validated.expires_in),
         ))
     }
 
@@ -91,7 +94,7 @@ impl TwitchToken for UserToken {
 
     fn token(&self) -> &AccessToken { &self.access_token }
 
-    fn login(&self) -> Option<&str> { self.login.as_deref() }
+    fn login(&self) -> Option<&str> { Some(&self.login) }
 
     async fn refresh_token<RE, C, F>(
         &mut self,
@@ -119,11 +122,13 @@ impl TwitchToken for UserToken {
         }
     }
 
-    fn expires_in(&self) -> Option<std::time::Duration> {
-        self.expires_in.map(|e| e - self.struct_created.elapsed())
+    fn expires_in(&self) -> std::time::Duration {
+        self.expires_in
+            .checked_sub(self.struct_created.elapsed())
+            .unwrap_or_default()
     }
 
-    fn scopes(&self) -> Option<&[Scope]> { Some(self.scopes.as_slice()) }
+    fn scopes(&self) -> &[Scope] { self.scopes.as_slice() }
 }
 
 /// Builder for [OAuth authorization code flow](https://dev.twitch.tv/docs/authentication/getting-tokens-oauth/#oauth-authorization-code-flow)
@@ -161,6 +166,15 @@ impl UserTokenBuilder {
             client_secret,
         })
     }
+
+    /// Add scopes to the request
+    pub fn set_scopes(mut self, scopes: Vec<Scope>) -> Self {
+        self.scopes = scopes;
+        self
+    }
+
+    /// Add a single scope to request
+    pub fn add_scope(&mut self, scope: Scope) { self.scopes.push(scope); }
 
     /// Enable or disable function to make the user able to switch accounts if needed.
     pub fn force_verify(mut self, b: bool) -> Self {